feature: Create remaining landing page components
