+++ src/pages/index.tsx
@@ -1,250 +1,22 @@
-import { ArrowTopRightOnSquareIcon } from '@heroicons/react/20/solid';
 import LandingPageLayout from '../layouts/LandingPageLayout';
 import Hero from '@/ui/Hero';
 import FeaturedScreenshot from '@/ui/FeaturedScreenshot';
 import FeaturedGrid from '@/ui/FeaturedGrid';
-
-const blogPosts = [
-  {
-    id: 1,
-    title: 'Boost your conversion rate',
-    href: '#',
-    date: 'Mar 16, 2020',
-    datetime: '2020-03-16',
-    category: { name: 'Article', href: '#' },
-    imageUrl:
-      'https://images.unsplash.com/photo-1496128858413-b36217c2ce36?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=1679&q=80',
-    preview:
-      'Lorem ipsum dolor sit amet consectetur adipisicing elit. Architecto accusantium praesentium eius, ut atque fuga culpa, similique sequi cum eos quis dolorum.',
-    author: {
-      name: 'Roel Aufderehar',
-      imageUrl:
-        'https://images.unsplash.com/photo-1472099645785-5658abf4ff4e?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80',
-      href: '#',
-    },
-    readingLength: '6 min',
-  },
-  {
-    id: 2,
-    title: 'How to use search engine optimization to drive sales',
-    href: '#',
-    date: 'Mar 10, 2020',
-    datetime: '2020-03-10',
-    category: { name: 'Video', href: '#' },
-    imageUrl:
-      'https://images.unsplash.com/photo-1547586696-ea22b4d4235d?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=1679&q=80',
-    preview:
-      'Lorem ipsum dolor sit amet consectetur adipisicing elit. Velit facilis asperiores porro quaerat doloribus, eveniet dolore. Adipisci tempora aut inventore optio animi., tempore temporibus quo laudantium.',
-    author: {
-      name: 'Brenna Goyette',
-      imageUrl:
-        'https://images.unsplash.com/photo-1550525811-e5869dd03032?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80',
-      href: '#',
-    },
-    readingLength: '4 min',
-  },
-  {
-    id: 3,
-    title: 'Improve your customer experience',
-    href: '#',
-    date: 'Feb 12, 2020',
-    datetime: '2020-02-12',
-    category: { name: 'Case Study', href: '#' },
-    imageUrl:
-      'https://images.unsplash.com/photo-1492724441997-5dc865305da7?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=1679&q=80',
-    preview:
-      'Lorem ipsum dolor sit amet consectetur adipisicing elit. Sint harum rerum voluptatem quo recusandae magni placeat saepe molestiae, sed excepturi cumque corporis perferendis hic.',
-    author: {
-      name: 'Daniela Metz',
-      imageUrl:
-        'https://images.unsplash.com/photo-1487412720507-e7ab37603c6f?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80',
-      href: '#',
-    },
-    readingLength: '11 min',
-  },
-];
+import Testimonial from '@/ui/Testimonial';
+import Blog from '@/ui/Blog';
+import CallToAction from '@/ui/CallToAction';
 
 export default function Home() {
   return (
-    <div className='bg-white'>
-      <div className='relative overflow-hidden'>
-        <LandingPageLayout>
-          <main>
-            <Hero />
-            <FeaturedScreenshot />
-            <FeaturedGrid />
-
-            {/* Testimonial section */}
-            <div className='pb-16 bg-gradient-to-r from-teal-500 to-cyan-600 lg:relative lg:z-10 lg:pb-0'>
-              <div className='lg:mx-auto lg:grid lg:max-w-7xl lg:grid-cols-3 lg:gap-8 lg:px-8'>
-                <div className='relative lg:-my-8'>
-                  <div
-                    aria-hidden='true'
-                    className='absolute inset-x-0 top-0 bg-white h-1/2 lg:hidden'
-                  />
-                  <div className='max-w-md px-6 mx-auto sm:max-w-3xl lg:h-full lg:p-0'>
-                    <div className='overflow-hidden shadow-xl aspect-w-10 aspect-h-6 rounded-xl sm:aspect-w-16 sm:aspect-h-7 lg:aspect-none lg:h-full'>
-                      <img
-                        className='object-cover lg:h-full lg:w-full'
-                        src='https://images.unsplash.com/photo-1520333789090-1afc82db536a?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=2102&q=80'
-                        alt=''
-                      />
-                    </div>
-                  </div>
-                </div>
-                <div className='mt-12 lg:col-span-2 lg:m-0 lg:pl-8'>
-                  <div className='max-w-md px-6 mx-auto sm:max-w-2xl lg:max-w-none lg:px-0 lg:py-20'>
-                    <blockquote>
-                      <div>
-                        <svg
-                          className='w-12 h-12 text-white opacity-25'
-                          fill='currentColor'
-                          viewBox='0 0 32 32'
-                          aria-hidden='true'>
-                          <path d='M9.352 4C4.456 7.456 1 13.12 1 19.36c0 5.088 3.072 8.064 6.624 8.064 3.36 0 5.856-2.688 5.856-5.856 0-3.168-2.208-5.472-5.088-5.472-.576 0-1.344.096-1.536.192.48-3.264 3.552-7.104 6.624-9.024L9.352 4zm16.512 0c-4.8 3.456-8.256 9.12-8.256 15.36 0 5.088 3.072 8.064 6.624 8.064 3.264 0 5.856-2.688 5.856-5.856 0-3.168-2.304-5.472-5.184-5.472-.576 0-1.248.096-1.44.192.48-3.264 3.456-7.104 6.528-9.024L25.864 4z' />
-                        </svg>
-                        <p className='mt-6 text-2xl font-medium text-white'>
-                          Lorem ipsum dolor sit amet, consectetur adipiscing
-                          elit. Sed urna nulla vitae laoreet augue. Amet feugiat
-                          est integer dolor auctor adipiscing nunc urna, sit.
-                        </p>
-                      </div>
-                      <footer className='mt-6'>
-                        <p className='text-base font-medium text-white'>
-                          Judith Black
-                        </p>
-                        <p className='text-base font-medium text-cyan-100'>
-                          CEO at PureInsights
-                        </p>
-                      </footer>
-                    </blockquote>
-                  </div>
-                </div>
-              </div>
-            </div>
-
-            {/* Blog section */}
-            <div className='relative py-16 bg-gray-50 sm:py-24 lg:py-32'>
-              <div className='relative'>
-                <div className='max-w-md px-6 mx-auto text-center sm:max-w-3xl lg:max-w-7xl lg:px-8'>
-                  <h2 className='text-lg font-semibold text-cyan-600'>Learn</h2>
-                  <p className='mt-2 text-3xl font-bold tracking-tight text-gray-900 sm:text-4xl'>
-                    Helpful Resources
-                  </p>
-                  <p className='mx-auto mt-5 text-xl text-gray-500 max-w-prose'>
-                    Phasellus lorem quam molestie id quisque diam aenean nulla
-                    in. Accumsan in quis quis nunc, ullamcorper malesuada.
-                    Eleifend condimentum id viverra nulla.
-                  </p>
-                </div>
-                <div className='grid max-w-md gap-8 px-6 mx-auto mt-12 sm:max-w-lg lg:max-w-7xl lg:grid-cols-3 lg:px-8'>
-                  {blogPosts.map((post) => (
-                    <div
-                      key={post.id}
-                      className='flex flex-col overflow-hidden rounded-lg shadow-lg'>
-                      <div className='flex-shrink-0'>
-                        <img
-                          className='object-cover w-full h-48'
-                          src={post.imageUrl}
-                          alt=''
-                        />
-                      </div>
-                      <div className='flex flex-col justify-between flex-1 p-6 bg-white'>
-                        <div className='flex-1'>
-                          <p className='text-sm font-medium text-cyan-600'>
-                            <a
-                              href={post.category.href}
-                              className='hover:underline'>
-                              {post.category.name}
-                            </a>
-                          </p>
-                          <a href={post.href} className='block mt-2'>
-                            <p className='text-xl font-semibold text-gray-900'>
-                              {post.title}
-                            </p>
-                            <p className='mt-3 text-base text-gray-500'>
-                              {post.preview}
-                            </p>
-                          </a>
-                        </div>
-                        <div className='flex items-center mt-6'>
-                          <div className='flex-shrink-0'>
-                            <a href={post.author.href}>
-                              <img
-                                className='w-10 h-10 rounded-full'
-                                src={post.author.imageUrl}
-                                alt={post.author.name}
-                              />
-                            </a>
-                          </div>
-                          <div className='ml-3'>
-                            <p className='text-sm font-medium text-gray-900'>
-                              <a
-                                href={post.author.href}
-                                className='hover:underline'>
-                                {post.author.name}
-                              </a>
-                            </p>
-                            <div className='flex space-x-1 text-sm text-gray-500'>
-                              <time dateTime={post.datetime}>{post.date}</time>
-                              <span aria-hidden='true'>&middot;</span>
-                              <span>{post.readingLength} read</span>
-                            </div>
-                          </div>
-                        </div>
-                      </div>
-                    </div>
-                  ))}
-                </div>
-              </div>
-            </div>
-
-            {/* CTA Section */}
-            <div className='relative bg-gray-900'>
-              <div className='relative h-56 bg-indigo-600 sm:h-72 md:absolute md:left-0 md:h-full md:w-1/2'>
-                <img
-                  className='object-cover w-full h-full'
-                  src='https://images.unsplash.com/photo-1525130413817-d45c1d127c42?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=1920&q=60&sat=-100'
-                  alt=''
-                />
-                <div
-                  aria-hidden='true'
-                  className='absolute inset-0 bg-gradient-to-r from-teal-500 to-cyan-600 mix-blend-multiply'
-                />
-              </div>
-              <div className='relative max-w-md px-6 py-12 mx-auto sm:max-w-7xl sm:py-20 md:py-28 lg:px-8 lg:py-32'>
-                <div className='md:ml-auto md:w-1/2 md:pl-10'>
-                  <h2 className='text-lg font-semibold text-gray-300'>
-                    Award winning support
-                  </h2>
-                  <p className='mt-2 text-3xl font-bold tracking-tight text-white sm:text-4xl'>
-                    We’re here to help
-                  </p>
-                  <p className='mt-3 text-lg text-gray-300'>
-                    Lorem ipsum dolor sit amet, consectetur adipiscing elit. Et,
-                    egestas tempus tellus etiam sed. Quam a scelerisque amet
-                    ullamcorper eu enim et fermentum, augue. Aliquet amet
-                    volutpat quisque ut interdum tincidunt duis.
-                  </p>
-                  <div className='mt-8'>
-                    <div className='inline-flex rounded-md shadow'>
-                      <a
-                        href='#'
-                        className='inline-flex items-center justify-center px-5 py-3 text-base font-medium text-gray-900 bg-white border border-transparent rounded-md hover:bg-gray-50'>
-                        Visit the help center
-                        <ArrowTopRightOnSquareIcon
-                          className='w-5 h-5 ml-3 -mr-1 text-gray-400'
-                          aria-hidden='true'
-                        />
-                      </a>
-                    </div>
-                  </div>
-                </div>
-              </div>
-            </div>
-          </main>
-        </LandingPageLayout>
-      </div>
-    </div>
+    <LandingPageLayout>
+      <main>
+        <Hero />
+        <FeaturedScreenshot />
+        <FeaturedGrid />
+        <Testimonial />
+        <Blog />
+        <CallToAction />
+      </main>
+    </LandingPageLayout>
   );
 }
